+++ src/cz_cpp/cz_Define.h
@@ -17,7 +17,25 @@
  * @brief  CubeZ Definition Header
  */
 
+#ifndef DISABLE_MPI
 #include <CB_Define.h>
+#else
+#include "CB_Define_stub.h"
+#endif
+
+
+// precision
+#ifdef _REAL_IS_DOUBLE_
+#define REAL_TYPE double
+#else
+// /** 実数型の指定
+//  * - デフォルトでは、REAL_TYPE=float
+//   * - コンパイル時オプション-D_REAL_IS_DOUBLE_を付与することで
+//    *   REAL_TYPE=doubleになる
+//     */
+#define REAL_TYPE float
+#endif
+
 
 #ifdef DISABLE_MPI
   typedef int MPI_Op;